kubernetes-9-prometheus/app/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.1
Flask-Migrate==3.1.0
prometheus-client==0.7.1